# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reviews": {
      "post": {
        "operationId": "ReviewsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created"
          },
          "400": {
            "description": "Booking not completed, duplicate review, or invalid service/provider"
          },
          "403": {
            "description": "Forbidden - customer only or not own booking"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Create a review for a completed booking service",
        "tags": [
          "reviews"
        ]
      },
      "get": {
        "operationId": "ReviewsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Reviews returned"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "List all reviews for operational dashboards",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/booking/{bookingId}": {
      "get": {
        "operationId": "ReviewsController_findByBooking",
        "parameters": [
          {
            "name": "bookingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking reviews returned"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "List reviews for a booking",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/provider/{spId}": {
      "get": {
        "operationId": "ReviewsController_findByProvider",
        "parameters": [
          {
            "name": "spId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider reviews returned"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "List reviews for a service provider",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/customer/{customerId}": {
      "get": {
        "operationId": "ReviewsController_findByCustomer",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer reviews returned"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "List reviews written by a customer",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/service/{serviceId}": {
      "get": {
        "operationId": "ReviewsController_findByService",
        "parameters": [
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service reviews returned"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "List reviews for a service",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/{id}": {
      "get": {
        "operationId": "ReviewsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review returned"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Review not found"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get review by ID",
        "tags": [
          "reviews"
        ]
      },
      "patch": {
        "operationId": "ReviewsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review updated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Review not found"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Update a review",
        "tags": [
          "reviews"
        ]
      },
      "delete": {
        "operationId": "ReviewsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review deleted"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Review not found"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Delete a review",
        "tags": [
          "reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "TatkuUnited API",
    "description": "Service platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste your JWT token (without the Bearer prefix)"
      },
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role header required by RolesGuard. Allowed values: super_user, region_manager, service_provider, customer"
      }
    },
    "schemas": {
...
      "CreateReviewDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateReviewDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}